+++ app/lib/ui/components/badge/badge_demo_screen.dart
@@ -13,7 +13,6 @@
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:ouds_core/components/badge/internal/ouds_badge_status_modifier.dart';
 import 'package:ouds_core/components/badge/ouds_badge.dart';
 import 'package:ouds_flutter_demo/l10n/app_localizations.dart';
 import 'package:ouds_flutter_demo/main_app_bar.dart';
@@ -137,13 +136,24 @@ class _BadgeDemoState extends State<_BadgeDemo> {
           child: Row(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
-              OudsBadge(
-                label: BadgeCustomizationUtils.getType(customizationState!.selectedType) == BadgeEnumType.count ? BadgeCustomizationUtils.getNumberText(customizationState!) : null,
-                icon: BadgeCustomizationUtils.getIcon(customizationState, themeController),
-                size: BadgeCustomizationUtils.getSize(customizationState!.selectedState),
-                status: BadgeCustomizationUtils.getStatus(customizationState!.selectedStatus),
-                enabled: customizationState!.hasEnabled,
-                semanticsLabel: BadgeCustomizationUtils().getSemanticLabel(context, customizationState!),
+              customizationState!.selectedType == BadgeEnumType.count
+                  ? OudsBadge.count(
+                  label: customizationState!.selectedType == BadgeEnumType.count ? BadgeCustomizationUtils.getNumberText(customizationState!) : null,
+                  size: BadgeCustomizationUtils.getSize(customizationState!.selectedState),
+                  status: BadgeCustomizationUtils.getStatus(customizationState!.selectedStatus),
+                  enabled: customizationState!.hasEnabled,
+                  semanticsLabel: BadgeCustomizationUtils.getSemanticLabel(context, customizationState!),
+              ) : customizationState!.selectedType == BadgeEnumType.icon
+                  ? OudsBadge.icon(
+                  size: BadgeCustomizationUtils.getSize(customizationState!.selectedState),
+                  enabled: customizationState!.hasEnabled,
+                  semanticsLabel: BadgeCustomizationUtils.getSemanticLabel(context, customizationState!),
+                  badgeIconStatus: BadgeCustomizationUtils.getBadgeStatus(context,customizationState!, themeController)
+              ) :  OudsBadge.standard(
+                  size: BadgeCustomizationUtils.getSize(customizationState!.selectedState),
+                  status: BadgeCustomizationUtils.getStatus(customizationState!.selectedStatus),
+                  enabled: customizationState!.hasEnabled,
+                  semanticsLabel: BadgeCustomizationUtils.getSemanticLabel(context, customizationState!),
               )
             ],
           ),
@@ -181,11 +191,11 @@ class _CustomizationContentState extends State<_CustomizationContent> {
   @override
   Widget build(BuildContext context) {
     final BadgeCustomizationState? customizationState = BadgeCustomization.of(context);
-    final badgeStatusModifier = OudsBadgeStatusModifier(context);
     final theme = OudsTheme.of(context).spaceScheme(context);
     var status = customizationState!.statusState.list;
     var size = customizationState.sizeState.list;
     var style = customizationState.typeState.list;
+    final themeController = Provider.of<ThemeController>(context, listen: true);
 
     return CustomizableSection(
       children: [
@@ -243,14 +253,18 @@ class _CustomizationContentState extends State<_CustomizationContent> {
                   width: theme.paddingBlockMedium,
                   height: theme.paddingBlockMedium,
                   decoration: BoxDecoration(
-                    color: badgeStatusModifier.getStatusColor(BadgeCustomizationUtils.getStatus(status), true),
+                    color: BadgeCustomizationUtils.getStatusColor(
+                      context,
+                        BadgeCustomizationUtils.getStatus(status),
+                        BadgeCustomizationUtils.getBadgeStatus(context, customizationState, themeController),
+                        true),
                     shape: BoxShape.rectangle,
                   ),
                 );
           }).toList(),
         ),
         CustomizableTextField(
-          fieldEnable: BadgeCustomizationUtils.getType(customizationState.selectedType) == BadgeEnumType.count,
+          fieldEnable: customizationState.selectedType == BadgeEnumType.count,
           keyboardType: TextInputType.number,
           title: context.l10n.app_components_badge_count_label,
           text: customizationState.countText,